Priestley's Chart of Biography Data

These are several datasets from the Joseph Priestley's *A Chart of Biography*, which is the first known use of a visual timeline to display historical data. The data consist of the lives of approximately 2,000 individuals between 1200 B.C.E. and about 1800 C.E. The *Chart of Biography* was published in numerous editions between 1764 and 1815. This package provides the data associated with the 1st (1764) and 74th (1778) editions.x

Biographies

Specimen

Format

An object of class tbl_df (inherits from tbl, data.frame) with 2418 rows and 26 columns.

Details

These datasets comprise the biographical data in Priestley's *Chart of Biography*.